feat: stabilize api

+++ tako-streams/src/file_stream.rs
@@ -454,7 +454,7 @@ fn epoch_days_to_ymd(days: i64) -> (i64, i64, i64) {
   // Civil from days since 1970-01-01 — Howard Hinnant algorithm.
   let z = days + 719468;
   let era = if z >= 0 { z } else { z - 146096 } / 146097;
-  let doe = (z - era * 146097) as i64;
+  let doe = z - era * 146097;
   let yoe = (doe - doe / 1460 + doe / 36524 - doe / 146096) / 365;
   let y = yoe + era * 400;
   let doy = doe - (365 * yoe + yoe / 4 - yoe / 100);
